Break out firebase logic
